Decouple websocket service from streamlabels service allowing multple listeners
