+++ app/services/streamlabels/index.ts
@@ -3,7 +3,8 @@ import { UserService } from 'services/user';
 import { Inject } from 'util/injector';
 import { HostsService } from 'services/hosts';
 import { handleErrors, authorizedHeaders } from 'util/requests';
-import io from 'socket.io-client';
+import { WebsocketService, TSocketEvent } from 'services/websocket';
+// import io from 'socket.io-client';
 import uuid from 'uuid/v4';
 import fs from 'fs';
 import path from 'path';
@@ -63,42 +64,8 @@ interface ITrains {
   follow: ITrainInfo;
 }
 
+
 type TTrainType = 'donation' | 'follow' | 'subscription';
-
-type TSocketEvent =
-  IStreamlabelsSocketEvent |
-  IDonationSocketEvent |
-  IFollowSocketEvent |
-  ISubscriptionSocketEvent;
-
-interface IStreamlabelsSocketEvent {
-  type: 'streamlabels';
-  message: {
-    data: Dictionary<string>;
-  };
-}
-
-interface IDonationSocketEvent {
-  type: 'donation';
-  message: {
-    name: string;
-    amount: string;
-  }[];
-}
-
-interface IFollowSocketEvent {
-  type: 'follow';
-  message: {
-    name: string;
-  }[];
-}
-
-interface ISubscriptionSocketEvent {
-  type: 'subscription';
-  message: {
-    name: string;
-  }[];
-}
 
 
 function isDonationTrain(train: ITrainInfo | IDonationTrainInfo): train is IDonationTrainInfo {
@@ -110,6 +77,7 @@ export class StreamlabelsService extends Service {
 
   @Inject() userService: UserService;
   @Inject() hostsService: HostsService;
+  @Inject() websocketService: WebsocketService;
 
 
   /**
@@ -180,7 +148,6 @@ export class StreamlabelsService extends Service {
   onUserLogin() {
     this.fetchInitialData();
     this.fetchSettings();
-    this.initSocketConnection();
   }
 
 
@@ -321,36 +288,7 @@ export class StreamlabelsService extends Service {
 
 
   private initSocketConnection(): void {
-    if (!this.userService.isLoggedIn()) {
-      console.warn('User must be logged in for streamlabels socket connection');
-      return;
-    }
-
-    if (this.socket) {
-      this.socket.disconnect();
-    }
-
-    const url = `https://${this.hostsService.streamlabs}/api/v5/slobs/socket-token`;
-    const headers = authorizedHeaders(this.userService.apiToken);
-    const request = new Request(url, { headers });
-
-    fetch(request)
-      .then(handleErrors)
-      .then(response => response.json())
-      .then(json => json.socket_token)
-      .then(token => {
-        const url = `https://aws-io.${this.hostsService.streamlabs}?token=${token}`;
-        this.socket = io(url, { transports: ['websocket'] });
-
-        // These are useful for debugging
-        this.socket.on('connect', () => this.log('Connection Opened'));
-        this.socket.on('connect_error', (e: any) => this.log('Connection Error', e));
-        this.socket.on('connect_timeout', () => this.log('Connection Timeout'));
-        this.socket.on('error', () => this.log('Error'));
-        this.socket.on('disconnect', () => this.log('Connection Closed'));
-
-        this.socket.on('event', (e: any) => this.onSocketEvent(e));
-      });
+    this.websocketService.socketEvent.subscribe(e => this.onSocketEvent(e));
   }
